Split up uptimerobot
